fix：意外启动s4u
