+++ src/chat/mai_thinking/mai_think.py
@@ -3,7 +3,7 @@ import time
 from src.chat.utils.prompt_builder import Prompt, global_prompt_manager
 from src.llm_models.utils_model import LLMRequest
 from src.config.config import global_config
-from src.chat.message_receive.message import MessageSending, MessageRecv, MessageRecvS4U
+from src.chat.message_receive.message import MessageRecvS4U
 from src.mais4u.mais4u_chat.s4u_msg_processor import S4UMessageProcessor
 from src.mais4u.mais4u_chat.internal_manager import internal_manager
 from src.common.logger import get_logger
