+++ src/mais4u/constant_s4u.py
@@ -1,0 +1,1 @@
+ENABLE_THINKING = False
